[Backend/writerspage] I created WriterController and moved all writer's page functions.
