+++ app/Http/Controllers/UserController.php
@@ -20,26 +20,4 @@ class UserController extends Controller
         $this->user = $user;
         $this->post = $post;
     }
-
-    private function getWritersPosts ($page = 1, $perPage = 9, $id) 
-    {
-        $writers_posts = $this->post->findOrFail($id)->
-        take(30)->get();
-        $writersRecentlyItems = $writers_posts->slice(($page - 1) * $perPage,$perPage)->all();
-        $writersPaginatedItem = new LengthAwarePaginator($writersRecentlyItems,count($writers_posts),$perPage,$page,['path' => LengthAwarePaginator::resolveCurrentPath()]);
-
-        $writersPaginatedItem->withPath('recently');
-
-        return $writersPaginatedItem;
-    }
-
-    public function writer(Request $request, $id) {
-        $writer = $this->user->findOrFail($id);
-        $page = $request->input('page',1);
-        $perPage = $request->input('perPage',9);
-
-        $writer_recently = $this->getWritersPosts($page,$perPage,$id);
-
-        return view('writers.writer',compact('writer','writer_recently'));
-    }
 }
